Move types caching to local storage to speed up initial load
