+++ PokeAutobuilder/Shared/MainLayout.razor
@@ -4,7 +4,7 @@
 @using PokeAutobuilder.Source.Services;
 @using PokemonDataModel
 
-@inject ProfileService ProfileService
+@inject ProfileService Profile
 @inject SessionService Session
 @inject PokeApiService PokeApiService
 @inject IAnalytics GlobalTracking
@@ -136,7 +136,7 @@
 
             // initialising global variables used across the app
             // load preferences first before the slow ones
-            Preferences prefs = await ProfileService.GetPreferencesAsync();
+            Preferences prefs = await Profile.GetPreferencesAsync();
             _isDarkMode = prefs.DarkMode;
 
             StateHasChanged();
@@ -144,9 +144,9 @@
             // session contains information about pokemon types that profile will use
             // to load pokemon. Therefore session must be done first!
             await Session.LoadSessionStorage();
-            await ProfileService.LoadProfileStorage();        
+            await Profile.LoadProfileStorage();        
 
-            if (Session.AllTypes.Count > 0)
+            if (Profile.AllTypes.Count > 0)
             {
                 _state = Init.Done;
             }
@@ -158,7 +158,7 @@
             }
 
             SearchLocations.Add(new SmartPokedex("National Pokédex", new NamedApiResource<Pokedex> { Name = "national", Url = "https://pokeapi.co/api/v2/pokedex/1/" }));
-            SearchLocations.Add(ProfileService.PokemonStorage.Boxes[0]);
+            SearchLocations.Add(Profile.PokemonStorage.Boxes[0]);
             SearchLocations.Add(new SmartPokedex("Red & Blue", new NamedApiResource<VersionGroup> { Name = "red-blue", Url = "https://pokeapi.co/api/v2/version-group/1/" }));
             SearchLocations.Add(new SmartPokedex("Yellow", new NamedApiResource<VersionGroup> { Name = "yellow", Url = "https://pokeapi.co/api/v2/version-group/2/" }));
             SearchLocations.Add(new SmartPokedex("Gold & Silver", new NamedApiResource<VersionGroup> { Name = "gold-silver", Url = "https://pokeapi.co/api/v2/version-group/3/" }));
@@ -192,7 +192,7 @@
 
     void OnChangeDarkMode()
     {
-        ProfileService?.SetDarkModeAsync(_IsDarkMode);
+        Profile?.SetDarkModeAsync(_IsDarkMode);
     }
 
     // nav menu
